+++ src/components/member/PaymentRequestDetail.vue
@@ -1,21 +1,6 @@
 <template>
   <div class="max-width">
-    <h3>금액 상세내역</h3>
-    <hr />
-    <table class="table">
-      <thead>
-        <tr>
-          <th scope="col">품목</th>
-          <th scope="col">가격</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr v-for="(item, index) in checklist" :key="index">
-          <td>{{ item.laundry.name }}</td>
-          <td>{{ item.laundry.laundryPrice.price }}</td>
-        </tr>
-      </tbody>
-    </table>
+    <CheckTable :id="confirmId" v-if="showCheck" />
 
     <button class="btn btn-secondary btn-sm" @click="goback">돌아가기</button>
     <button class="btn btn-primary btn-sm" @click="showModal">결제하기</button>
@@ -25,15 +10,21 @@
 <script>
 import router from '@/router/router'
 import { useRoute } from 'vue-router'
-import { getPrDetail, getCheckList } from '@/api'
 import { onMounted, ref } from 'vue'
+import { getPrDetail } from '@/api'
 import { failToast, paymentModal } from '@/sweetAlert'
+import CheckTable from './CheckTable.vue'
+
 export default {
+  components: {
+    CheckTable,
+  },
   setup() {
     const route = useRoute()
     const prId = route.params.id
+    const confirmId = ref(-1)
     const pr = ref({})
-    const checklist = ref({})
+    const showCheck = ref(false)
 
     onMounted(() => {
       getPrDetail(prId)
@@ -41,23 +32,13 @@ export default {
           console.log(res.data)
           pr.value = res.data
 
-          callCheckAPI(pr.value.confirm.confirmId)
+          confirmId.value = pr.value.confirm.confirmId
+          showCheck.value = true
         })
         .catch(() => {
           failToast('결제요청서 로딩에 실패하였습니다.')
         })
     })
-
-    const callCheckAPI = (id) => {
-      getCheckList(id)
-        .then((res) => {
-          console.log(res.data)
-          checklist.value = res.data
-        })
-        .catch(() => {
-          failToast('검수내역 로딩에 실패하였습니다.')
-        })
-    }
 
     const showModal = () => {
       paymentModal(prId)
@@ -72,9 +53,9 @@ export default {
       route,
       prId,
       pr,
-      callCheckAPI,
-      checklist,
       showModal,
+      confirmId,
+      showCheck,
     }
   },
 }
